Refactor logic shared across table views
