+++ mlflow/server/js/src/components/ExperimentViewUtil.js
@@ -135,6 +135,49 @@ export default class ExperimentViewUtil {
     />;
   }
 
+
+  static getUnbaggedMetricCell(i, metricKey, metricsMap, metricRanges) {
+    const className = (i === 0 ? "left-border" : "") + " run-table-container";
+    const keyName = "metric-" + metricKey;
+    if (metricsMap[metricKey]) {
+      const metric = metricsMap[metricKey].getValue();
+      const range = metricRanges[metricKey];
+      let fraction = 1.0;
+      if (range.max > range.min) {
+        fraction = (metric - range.min) / (range.max - range.min);
+      }
+      const percent = (fraction * 100) + "%";
+      return (
+        <td className={className} key={keyName}>
+          {/* We need the extra div because metric-filler-bg is inline-block */}
+          <div>
+            <div className="metric-filler-bg">
+              <div className="metric-filler-fg" style={{width: percent}}/>
+              <div className="metric-text">
+                {Utils.formatMetric(metric)}
+              </div>
+            </div>
+          </div>
+        </td>
+      );
+    }
+    return <td className={className} key={keyName}/>;
+  }
+
+  static getUnbaggedParamCell(i, paramKey, paramsMap) {
+    const className = (i === 0 ? "left-border" : "") + " run-table-container";
+    const keyName = "param-" + paramKey;
+    if (paramsMap[paramKey]) {
+      return <td className={className} key={keyName}>
+        <div>
+          {paramsMap[paramKey].getValue()}
+        </div>
+      </td>;
+    } else {
+      return <td className={className} key={keyName}/>;
+    }
+  }
+
   static isSortedBy(sortState, isMetric, isParam, key) {
     return (sortState.isMetric === isMetric && sortState.isParam === isParam
       && sortState.key === key);
